+++ packages/devtools/src/config/paths.ts
@@ -36,7 +36,7 @@ const libPackages = [
   'packages/tsconfig',
   'packages/useful-types',
   'packages/util',
-  'packages/use-parentsize-svg',
+  'packages/use-get-parent-size',
   'packages/hooks',
   'packages/use-mathjax',
   'packages/component-library',