+++ packages/website/CHANGELOG.md
@@ -35,7 +35,7 @@
 
 - Updated dependencies [undefined]
   - @cutting/hooks@4.25.0
-  - @cutting/use-parentsize-svg@1.1.0
+  - @cutting/use-get-parent-size@1.1.0
   - @cutting/component-library@4.23.4
   - @cutting/svg@4.25.1